Added default suggestion for data types in the frontend

+++ dbrepo-metadata-db/entities/src/main/java/at/tuwien/entities/database/table/columns/TableColumn.java
@@ -111,10 +111,10 @@ public class TableColumn implements Comparable<TableColumn> {
     private List<String> sets;
 
     @Column
-    private Long size;
+    private Integer size;
 
     @Column
-    private Long d;
+    private Integer d;
 
     @LastModifiedDate
     @Column(columnDefinition = "TIMESTAMP")
